Require permissions to use /ticket for creation

+++ src/commands/utility/ticket.js
@@ -1,11 +1,28 @@
-import { SlashCommandBuilder } from 'discord.js';
+import { SlashCommandBuilder, EmbedBuilder } from 'discord.js';
 import ticketSetupEmbed from '../../tickets/embeds/ticketSetup';
+import { supportRoleId } from '../../../config.json';
 
 module.exports = {
     data: new SlashCommandBuilder()
         .setName('ticket')
         .setDescription('Creates a new ticket setup.'),
     async execute(interaction) {
+        function errorEmbed(message) {
+            return new EmbedBuilder()
+                .setColor('#2f3137')
+                .setDescription(message);
+        }
+
+        if (!interaction.member.roles.cache.has(supportRoleId))
+            return interaction.reply({
+                embeds: [
+                    errorEmbed(
+                        'You do not have permission to use this command.'
+                    )
+                ],
+                ephemeral: true
+            });
+
         // Prevents there from being a reply sent.
         interaction.deferReply();
         interaction.deleteReply();
